splittng up candidate models into deeper file structure
