+++ hknweb/candidate/models/__init__.py
@@ -1,0 +1,19 @@
+from hknweb.candidate.models.activities import (
+    OffChallenge,
+    BitByteActivity,
+)
+from hknweb.candidate.models.announcement import Announcement
+
+from hknweb.candidate.models.requirements import (
+    CandidateForm,
+    CandidateFormDoneEntry,
+    CommitteeProject,
+    CommitteeProjectDoneEntry,
+    DuePayment,
+    DuePaymentPaidEntry,
+    RequirementBitByteActivity,
+    RequriementEvent,
+    RequirementHangout,
+    RequirementMandatory,
+    RequirementMergeRequirement,
+)
