+++ hknweb/candidate/models/activities/__init__.py
@@ -1,0 +1,2 @@
+from hknweb.candidate.models.activities.officer_challenge import OffChallenge
+from hknweb.candidate.models.activities.bit_byte_activity import BitByteActivity
