+++ hknweb/candidate/models/constants.py
@@ -1,0 +1,2 @@
+MAX_STRLEN = 150  # default max length for char fields
+MAX_TXTLEN = 2000  # default max length for text fields
