+++ hknweb/candidate/models/requirements/__init__.py
@@ -1,0 +1,11 @@
+from hknweb.candidate.models.requirements.candidate_forms import CandidateForm, CandidateFormDoneEntry
+
+from hknweb.candidate.models.requirements.payment import DuePayment, DuePaymentPaidEntry
+
+from hknweb.candidate.models.requirements.committee_project import CommitteeProject, CommitteeProjectDoneEntry
+
+from hknweb.candidate.models.requirements.event import RequriementEvent
+from hknweb.candidate.models.requirements.bit_byte_activity import RequirementBitByteActivity
+from hknweb.candidate.models.requirements.hangout import RequirementHangout
+from hknweb.candidate.models.requirements.mandatory import RequirementMandatory
+from hknweb.candidate.models.requirements.merge import RequirementMergeRequirement
