+++ hknweb/candidate/models/requirements/candidate_forms/__init__.py
@@ -1,0 +1,2 @@
+from hknweb.candidate.models.requirements.candidate_forms.candidate_form import CandidateForm
+from hknweb.candidate.models.requirements.candidate_forms.candidate_form_done_entry import CandidateFormDoneEntry
