+++ hknweb/candidate/models/requirements/committee_project/__init__.py
@@ -1,0 +1,2 @@
+from hknweb.candidate.models.requirements.committee_project.committee_project import CommitteeProject
+from hknweb.candidate.models.requirements.committee_project.committee_project_done_entry import CommitteeProjectDoneEntry
