+++ hknweb/candidate/models/requirements/payment/__init__.py
@@ -1,0 +1,2 @@
+from hknweb.candidate.models.requirements.payment.due_payment import DuePayment
+from hknweb.candidate.models.requirements.payment.due_payment_paid_entry import DuePaymentPaidEntry
